Add SinglePlayPage and remove PlayPage, also remove Navigation and add it to the Pages itself.
