+++ src/pages/HistoryPage.js
@@ -1,9 +1,11 @@
+import { Link } from 'react-router-dom/cjs/react-router-dom.min'
 import styled from 'styled-components/macro'
 import HistoryElement from '../components/HistoryElement'
 
 export default function HistoryPage({ history }) {
   return (
     <Container>
+      <Link to="/">&lt;-- start new</Link>
       <List>
         {history.map(({ playlistName, playerName, date, score, id }) => (
           <HistoryElement
@@ -20,7 +22,8 @@ export default function HistoryPage({ history }) {
 }
 
 const Container = styled.main`
-  padding: 2rem;
+  padding-top: 0.5rem;
+  padding: 0.5rem 1rem 0;
   height: 100vh;
   overflow-y: auto;
 `
@@ -29,5 +32,5 @@ const List = styled.ul`
   display: grid;
   gap: 1.5rem;
   list-style: none;
-  padding: 0;
+  padding: 0 1rem 1rem;
 `
